refactor instrumentation packages
