+++ src/main/java/rocks/inspectit/gepard/agent/instrumentation/InstrumentationManager.java
@@ -5,9 +5,10 @@ import java.lang.instrument.Instrumentation;
 import java.time.Duration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import rocks.inspectit.gepard.agent.instrumentation.filling.ClassDiscoveryService;
-import rocks.inspectit.gepard.agent.instrumentation.filling.ConfigurationReceiver;
-import rocks.inspectit.gepard.agent.instrumentation.processing.BatchInstrumenter;
+import rocks.inspectit.gepard.agent.instrumentation.cache.PendingClassesCache;
+import rocks.inspectit.gepard.agent.instrumentation.cache.input.ClassDiscoveryService;
+import rocks.inspectit.gepard.agent.instrumentation.cache.input.ConfigurationReceiver;
+import rocks.inspectit.gepard.agent.instrumentation.cache.process.BatchInstrumenter;
 import rocks.inspectit.gepard.agent.internal.configuration.observer.ConfigurationReceivedEvent;
 import rocks.inspectit.gepard.agent.internal.schedule.InspectitScheduler;
 import rocks.inspectit.gepard.agent.resolver.ConfigurationResolver;
